Convert ToDouble

+++ Program.cs
@@ -1,7 +1,7 @@
 ﻿Console.WriteLine("Inserisci il primo numero.");
-double num1 = Convert.ToInt32(Console.ReadLine());
+double num1 = Convert.ToDouble(Console.ReadLine());
 Console.WriteLine("Inserisci il secondo numero.");
-double num2 = Convert.ToInt32(Console.ReadLine());
+double num2 = Convert.ToDouble(Console.ReadLine());
 Console.WriteLine("Inserisci l'operatore tra +-*/");
 string? operatore = Console.ReadLine();
